+++ src/plugins/index.js
@@ -1,6 +1,15 @@
 import Vue from 'vue'
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { faEnvelope, faCalendarAlt, faCar, faHome, faMapMarked, faRss } from '@fortawesome/free-solid-svg-icons'
+import { faGithub, faLinkedinIn, faTwitter } from '@fortawesome/free-brands-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import Buefy from 'buefy'
 import VueHead from 'vue-head'
+
+library.add(faEnvelope, faCalendarAlt, faCar, faHome, faMapMarked, faRss)
+library.add(faGithub, faLinkedinIn, faTwitter)
+
+Vue.component('font-awesome-icon', FontAwesomeIcon)
 
 Vue.use(Buefy)
 Vue.use(VueHead)